reminder-py/requirements.txt
python-telegram-bot == 13.14
gspread == 5.5.0
oauth2client == 4.1.3
